fix(backend): BUG-BE-RPC-003 — migrations + pg-adapter sửa xong cho Postgres, bật lại
Audit + sửa dứt điểm 3 lớp lỗi phát hiện khi test thật với Postgres cục bộ (không chỉ tin code,
đã verify bằng docker run postgres:16-alpine + chạy MigrationRunner thật + đúng luồng
project.create đã crash trong production):

1. 7 file migration dùng SQL SQLite thuần (datetime('now'), strftime(), AUTOINCREMENT) — thêm
   backend/src/main/db/migrations/sql-dialect.ts (3 helper thuần hàm rẽ nhánh theo
   db.capabilities.dialect), áp dụng vào 0001/0002/0003/0004/0005/0008/0010.

2. pg-adapter.ts CHƯA BAO GIỜ dịch placeholder — toàn bộ codebase dùng `?` (quy ước SQLite),
   nhưng driver `pg` cần `$1, $2, $3`. Phát hiện qua MigrationRunner's chính nó
   (INSERT INTO schema_migrations) crash "syntax error at or near ','" — ảnh hưởng MỌI query
   trong toàn hệ thống, không chỉ migration. Sửa 1 chỗ: translatePlaceholders() trong
   pg-adapter.ts, bỏ qua `?` trong chuỗi literal.

3. Cột INTEGER lưu epoch-milliseconds (created_at, updated_at, expires_at, added_at, ...) tràn
   số ở Postgres (INTEGER 32-bit, Date.now() là 13 chữ số) — SQLite không lộ vì INTEGER affinity
   linh hoạt. Đổi ~30 cột timestamp sang BIGINT (an toàn cả 2 dialect, giữ nguyên cột nhỏ thật:
   port/enabled/priority/boolean).

Verify: dựng Postgres cục bộ, chạy MigrationRunner thật với ALL_MIGRATIONS (17 file) — sạch.
Test lại đúng luồng project.create (insert orca_users → orca_v5_projects →
orca_v5_project_members với FK) — thành công, không còn FOREIGN KEY constraint failed.
Test mới: sql-dialect.test.ts (9 case), pg-adapter.test.ts (6 case). Backend suite 139/139 pass.
gitnexus detect_changes: risk LOW, 0 affected_processes.

Đã bật lại ORCA_DB_URL + depends_on: postgres trong cả 3 file compose (tắt tạm ở commit trước để
phục hồi dịch vụ). Chi tiết đầy đủ: docs/guides/postgres-shared-database-design.md

+++ backend/src/main/db/migrations/0016_team_project_sharing_task_exec.ts
@@ -28,8 +28,8 @@ export const migration0016TeamProjectSharingTaskExec: Migration = {
       CREATE TABLE IF NOT EXISTS orca_teams (
         id         TEXT    PRIMARY KEY,
         name       TEXT    NOT NULL,
-        created_at INTEGER NOT NULL,
-        updated_at INTEGER NOT NULL
+        created_at BIGINT NOT NULL,
+        updated_at BIGINT NOT NULL
       )
     `)
 
@@ -46,7 +46,7 @@ export const migration0016TeamProjectSharingTaskExec: Migration = {
         orca_project_id TEXT    NOT NULL REFERENCES orca_v5_projects(id) ON DELETE CASCADE,
         owner_user_id   TEXT    NOT NULL REFERENCES orca_users(id) ON DELETE CASCADE,
         project_id      TEXT    NOT NULL,
-        created_at      INTEGER NOT NULL,
+        created_at      BIGINT NOT NULL,
         PRIMARY KEY (orca_project_id, owner_user_id, project_id)
       )
     `)
